Formatting, documentation

+++ build-libmpdclient.sh
@@ -5,13 +5,22 @@ DEBDIR="/root/latestbuilds"
 MAINTAINER="root@stor.nyxi.eu"
 LOGFILE="/root/logbuilds"
 
+if [ "$1" == "" ]; then
+        echo "Must provide version to download!"
+        exit
+fi
+
+#####################
+# Cleanup, dl, untar
+#####################
 cd "$BUILDDIR"
 rm -R libmpdclient*
 wget http://www.musicpd.org/download/libmpdclient/2/libmpdclient-$1.tar.bz2
 tar xvf libmpdclient*tar.bz2
 rm libmpdclient*tar.bz2
 cd libmpdclient*
-./configure
+#####################
+
 #####################
 # Control files
 #####################
@@ -28,11 +37,18 @@ echo "Priority: Extra" >> tmp/DEBIAN/control
 echo "Filename: pool/main/libm/libmpdclient/libmpdclient2_$1_amd64.deb" >> tmp/DEBIAN/control
 echo "Description: cli client for the music player daemon (mpd)" >> tmp/DEBIAN/control
 chmod -R a-s tmp/DEBIAN
-make DESTDIR=$(pwd)/tmp install
 echo "#!/bin/sh" >> tmp/DEBIAN/postinst
 echo "ldconfig" >> tmp/DEBIAN/postinst
 chmod 755 tmp/DEBIAN/postinst
+#####################
+
+#####################
+# Build it!
+#####################
+./configure
+make DESTDIR=$(pwd)/tmp install
 dpkg-deb --build tmp "libmpdclient2_$1_amd64.deb"
 rm "$DEBDIR/libmpdclient2*deb"
 mv *deb "$DEBDIR/"
 echo "$(date) - libmpdclient2 $1 build ready" >> "$LOGFILE"
+#####################
