+++ build-mpc.sh
@@ -5,13 +5,22 @@ DEBDIR="/root/latestbuilds"
 MAINTAINER="root@stor.nyxi.eu"
 LOGFILE="/root/logbuilds"
 
+if [ "$1" == "" ]; then
+        echo "Must provide version to download!"
+        exit
+fi
+
+#####################
+# Cleanup, dl, untar
+#####################
 cd "$BUILDDIR"
 rm -R mpc*
 wget http://www.musicpd.org/download/mpc/0/mpc-$1.tar.bz2
 tar xvf mpc*tar.bz2
 rm mpc*tar.bz2
 cd mpc*
-./configure
+#####################
+
 #####################
 # Control files
 #####################
@@ -27,8 +36,15 @@ echo "Section: main" >> tmp/DEBIAN/control
 echo "Filename: pool/main/m/mpc/mpc_$1_amd64.deb" >> tmp/DEBIAN/control
 echo "Description: cli client for the music player daemon (mpd)" >> tmp/DEBIAN/control
 chmod -R a-s tmp/DEBIAN
+#####################
+
+#####################
+# Build it!
+#####################
+./configure
 make DESTDIR="$(pwd)"/tmp install
 dpkg-deb --build tmp "mpc_$1_amd64.deb"
 rm "$DEBDIR/mpc*deb" 2> /dev/null
 mv *deb "$DEBDIR/"
 echo "$(date) - mpc $1 build ready" >> "$LOGFILE"
+#####################
