+++ build-znc.sh
@@ -5,13 +5,22 @@ DEBDIR="/root/latestbuilds"
 MAINTAINER="root@stor.nyxi.eu"
 LOGFILE="/root/logbuilds"
 
+if [ "$1" == "" ]; then
+        echo "Must provide version to download!"
+        exit
+fi
+
+#####################
+# Cleanup, dl, untar
+#####################
 cd "$BUILDDIR"
 rm -R znc*
 wget http://znc.in/releases/znc-$1.tar.gz
 tar xvf znc*tar.gz
 rm znc*tar.gz
 cd znc*
-./configure
+#####################
+
 #####################
 # Control files
 #####################
@@ -27,11 +36,18 @@ echo "Section: main" >> tmp/DEBIAN/control
 echo "Filename: pool/main/z/znc/znc_$1_amd64.deb" >> tmp/DEBIAN/control
 echo "Description: znc is an IRC proxy (bouncer)." >> tmp/DEBIAN/control
 chmod -R a-s tmp/DEBIAN
-make DESTDIR="$(pwd)"/tmp install
 echo "#!/bin/sh" >> tmp/DEBIAN/postinst
 echo "ldconfig" >> tmp/DEBIAN/postinst
 chmod 755 tmp/DEBIAN/postinst
+#####################
+
+#####################
+# Build it!
+#####################
+./configure
+make DESTDIR="$(pwd)"/tmp install
 dpkg-deb --build tmp "znc_$1_amd64.deb"
 rm "$DEBDIR/znc*deb" 2> /dev/null
 mv *deb "$DEBDIR/"
 echo "$(date) - znc $1 build ready" >> "$LOGFILE"
+#####################
